backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary   # Postgres driver
python-dotenv     # Manage environment variables
alembic           # For DB migrations
pydantic          # (already a dependency of FastAPI, but useful to pin later)
